MCU Ram to module.

+++ src/gui/memory/mcumonitor.h
@@ -38,7 +38,7 @@ class MCUMonitor : public QDialog, private Ui::McuMonitor
 
         eMcu* m_processor;
 
-        uint8_t* m_statusReg;  // STATUS register
+        uint32_t* m_statusReg;  // STATUS register
 
         Watcher*  m_watcher;
         RamTable* m_ramTable;
