+++ src/gui/memory/memtable.h
@@ -56,7 +56,7 @@ class MemTable : public QWidget, private Ui::MemTable
         QTableWidgetItem* m_hoverItem;
 
         Memory* m_memory;
-        std::vector<uint64_t>* m_data;
+        std::vector<uint32_t>* m_data;
 };
 
 #endif
